Removed ParamStore ABC
