+++ entropylab/dashboard/pages/params/callbacks.py
@@ -1,7 +1,7 @@
 from dash import callback_context
 from dash.dependencies import Input, Output, State
 
-from entropylab.pipeline.api.param_store import ParamStore
+from entropylab import InProcessParamStore
 from entropylab.dashboard.pages.params.utils import (
     param_store_to_df,
     param_store_to_commits_df,
@@ -9,7 +9,7 @@ from entropylab.dashboard.pages.params.utils import (
 )
 
 
-def register_callbacks(app, param_store: ParamStore):
+def register_callbacks(app, param_store: InProcessParamStore):
     @app.callback(
         [
             Output("data-table", "data"),
